Use the local window object of the observed node
When observing resizeable elements rendered in a remote document via a portal, requestAnimationFrame calls can hang if they are not invoked in the local window of the node being observed. Most browsers will defer requestAnimationFrame callbacks that are called in background tabs or iframes until the tab or frame becomes visible again.